Mapping GET lists

+++ src/main/java/com/allissondev/api_gamelist/entities/GameList.java
@@ -10,22 +10,22 @@ public class GameList {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
 
     public GameList() {
     }
 
-    public GameList(int id, String name) {
+    public GameList(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
